Update humidifier.py
Added modes for AUTO and Quiet humidifier modes, the mode codes were added previously to the frigidaire library, not entirely sure why they were not added here.

Also not sure if MODE_SLEEP is best used for quiet, but that is what I have it set to for now and it at least doesn't error.