Fix breaking changes to frigidaire API
A new API broke the current version of this library, and was noticed
in 1/26/2024. This embeds a patched version of the frigidaire dependency
temporarily while I wait for a fix to get merged upstream. Additionally,
this introduces minor changes to accommodate the new API, as the
patched dependency was not able to maintain 100% API compatibility.